requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coverage==5.5
cycler==0.11.0
deprecation==2.1.0
dnspython==2.3.0
eyed3==0.9.7
filetype==1.2.0
Flask==2.2.3
fonttools==4.39.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
retrying==1.3.4
six==1.16.0
toml==0.10.2
urllib3==1.26.15
Werkzeug==2.2.3
wordcloud==1.8.2.2
ybc-commons==0.6.3
ybc-config==1.2.2
ybc-exception==1.2.1
ybc-face==5.1.15
zipp==3.15.0
